requirements.txt
setuptools>=63.3.0
requests>=2.28.1
ujson>=5.4.0
slovnet>=0.5.0
translate>=3.6.1
htmldate>=1.3.0
newspaper3k>=0.2.8
pandas>=1.4.2
numpy>=1.23.1